% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picMort-package.R
\docType{package}
\name{picMort-package}
\alias{picMort}
\alias{picMort-package}
\title{picMort: Reproducible Paediatric ICU Mortality Benchmark on PIC V1.1.0}
\description{
Tools and pipelines for the PICU mortality + LLM paper series (P1 baseline, P2 inductive-CI inference, P3 multi-modal fusion). Authoritative cohort specification, T+24h prediction-window feature extraction, PIM3 reconstruction, calibration and decision-curve evaluation. The package is the methodological substrate for the series; individual papers are drafted as vignettes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/max578/picMort}
  \item \url{https://max578.github.io/picMort}
  \item Report bugs at \url{https://github.com/max578/picMort/issues}
}

}
\author{
\strong{Maintainer}: Max Moldovan \email{max.moldovan@adelaide.edu.au} (\href{https://orcid.org/0000-0001-9680-8474}{ORCID})

Authors:
\itemize{
  \item Usman Iqbal [thesis advisor]
}

}
\keyword{internal}
